components/x11/mesa: build i915 driver instead of gallium driver
Revert to i915 dri driver. We can enable i915g gallium driver in future,
or use the crocus driver. For now we retain i915 as a legacy dri driver.

note: currently broken. Xorg log shows EE for no devices found. I read
this to be an error with libdrm as a result of our update in 2024.